backend/pyproject.toml
[tool.poetry]
name = "produtor-rural-techagro"
version = "0.1.0"
description = "Gerenciamento de Produtores Rurais com Clean Architecture e Django."
authors = ["Edson Bezerra <edcastanha@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0"
djangorestframework = "^3.15"
psycopg2-binary = "^2.9"
python-decouple = "^3.8" # Para gerenciar variáveis de ambiente
# pydantic = "^2.0" # Opcional, se usar Pydantic para DTOs
gunicorn = "^23.0.0"
drf-spectacular = "^0.28.0"
dj-database-url = "^3.0.0"
pytest-cov = "^6.2.1"
loguru = "^0.7.3"
mkdocs = "^1.6.1"
mkdocstrings = "^0.29.1"
mkdocstrings-python = "^1.16.12"
mkdocs-material = "^9.6.15"
djangorestframework-simplejwt = "^5.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.8"
pytest-asyncio = "^0.23"
pytest-mock = "^3.12"
ruff = "^0.4"
black = "^24.4"
mypy = "^1.10"
pytest-cov = "^6.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"